screen navigation fixed,continue button fixed
