+++ lib/Routes.dart
@@ -5,6 +5,7 @@ import 'package:needs_app/Screens/Forget_Screen/forgetscreen.dart';
 import 'package:needs_app/Screens/HomeScreen/homeScreen.dart';
 import 'package:needs_app/Screens/LoginPage/loginPage.dart';
 import 'package:needs_app/Screens/OTP_Screen/OTPScreen.dart';
+import 'package:needs_app/Screens/ProfilePage/profilePage.dart';
 import 'package:needs_app/Screens/SignUp_screen/signupscreen.dart';
 import 'package:needs_app/Screens/profile/profile.dart';
 import 'package:needs_app/Screens/splash/splashScreens.dart';
@@ -19,4 +20,5 @@ Map<String, WidgetBuilder> routes = {
   DetailsScreen.routeName: (context) => DetailsScreen(),
   CartScreen.routeName: (context) => CartScreen(),
   Profile.routeName: (context) => Profile(),
+  ProfilePage.routeName: (context) => ProfilePage(),
 };
