+++ lib/Screens/HomeScreen/homeScreen.dart
@@ -1,15 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:needs_app/sizeConfig.dart';
-
-import '../../BottomnavBar.dart';
+import '../../components/BottomnavBar.dart';
 import '../../Menustate.dart';
 import 'Body.dart';
 
 class HomeScreen extends StatelessWidget {
-  static const routeName = '/HomeScrene';
+  static const routeName = '/HomeScreen';
   @override
   Widget build(BuildContext context) {
-    SizeConfig().init(context);
     return Scaffold(
       bottomNavigationBar: BottomNavBar(
         selected: MenuState.homepage,
